# app/api/projects/[id]/domain/verify/route.ts
import { getAuthenticatedUser } from "@/lib/auth-helpers";
import { NextRequest, NextResponse } from "next/server";
import { withRateLimit } from "@/lib/rate-limit";
import { createClient } from "@/lib/supabase/server";
import dns from "dns";
import { promisify } from "util";

const resolveTxt = promisify(dns.resolveTxt);
const resolveCname = promisify(dns.resolveCname);

export async function POST(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> },
) {
  const limited = await withRateLimit(request, { limit: 20, window: 60_000 });
  if (limited) return limited;

  try {
    const { uid } = await getAuthenticatedUser(request);
    if (!uid) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    const { id } = await params;
    const supabase = await createClient();

    const { data: project, error } = await supabase
      .from("projects")
      .select("id, custom_domain, domain_verified")
      .eq("id", id)
      .single();

    if (error || !project) {
      return NextResponse.json({ error: "Project not found" }, { status: 404 });
    }

    if (!project.custom_domain) {
      return NextResponse.json(
        { error: "No custom domain configured" },
        { status: 400 },
      );
    }

    if (project.domain_verified) {
      return NextResponse.json({
        verified: true,
        message: "Domain is already verified",
      });
    }

    const domain = project.custom_domain;
    const expectedTxtRecord = `rux-verify=${id}`;

    let txtVerified = false;
    let cnameVerified = false;
    const errors: string[] = [];

    // Check TXT record for verification
    try {
      const txtRecords = await resolveTxt(`_rux.${domain}`);
      const flatRecords = txtRecords.flat();
      txtVerified = flatRecords.some((record) =>
        record.includes(expectedTxtRecord),
      );
      if (!txtVerified) {
        errors.push(
          `TXT record found but doesn't match. Expected: ${expectedTxtRecord}`,
        );
      }
    } catch (error: unknown) {
      const dnsError = error as { code?: string };
      if (dnsError.code === "ENODATA" || dnsError.code === "ENOTFOUND") {
        errors.push(`TXT record not found at _rux.${domain}`);
      } else {
        errors.push(
          `Failed to query TXT record: ${dnsError.code || "unknown error"}`,
        );
      }
    }

    // Check CNAME record
    try {
      const cnameRecords = await resolveCname(domain);
      cnameVerified = cnameRecords.some(
        (record) => record.toLowerCase() === "cname.rulxy.com",
      );
      if (!cnameVerified) {
        errors.push(
          `CNAME record found but points to wrong target. Expected: cname.rulxy.com, Got: ${cnameRecords[0]}`,
        );
      }
    } catch (error: unknown) {
      const dnsError = error as { code?: string };
      if (dnsError.code === "ENODATA" || dnsError.code === "ENOTFOUND") {
        errors.push(`CNAME record not found for ${domain}`);
      } else {
        errors.push(
          `Failed to query CNAME record: ${dnsError.code || "unknown error"}`,
        );
      }
    }

    const verified = txtVerified && cnameVerified;

    if (verified) {
      // Update project as verified using snake_case
      const { error: updateError } = await supabase
        .from("projects")
        .update({
          domain_verified: true,
          updated_at: new Date().toISOString(),
        })
        .eq("id", id);

      if (updateError) {
        console.error("Failed to update domain verification:", updateError);
        return NextResponse.json(
          { error: "Failed to update domain verification" },
          { status: 500 },
        );
      }

      return NextResponse.json({
        verified: true,
        message: "Domain verified successfully!",
        customDomainUrl: `https://${domain}`,
      });
    }

    return NextResponse.json({
      verified: false,
      txtVerified,
      cnameVerified,
      errors,
      message: "Domain verification failed. Please check your DNS settings.",
      requiredRecords: [
        {
          type: "CNAME",
          name: domain,
          value: "cname.rulxy.com",
          status: cnameVerified ? "ok" : "missing",
        },
        {
          type: "TXT",
          name: `_rux.${domain}`,
          value: expectedTxtRecord,
          status: txtVerified ? "ok" : "missing",
        },
      ],
    });
  } catch (error) {
    console.error("Failed to verify domain:", error);
    return NextResponse.json(
      { error: "Failed to verify domain" },
      { status: 500 },
    );
  }
}