# lib/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rulxy Platform API",
    "description": "Internal API documentation for the Rulxy platform. Covers project management, code generation, builds, proxy services, billing, and admin endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase JWT token"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RUX-API-Key",
        "description": "RUX API key for proxy services"
      }
    }
  },
...
  "paths": {
...
    "/api/projects/{id}/domain/verify": {
      "post": {
        "summary": "Verify custom domain DNS settings",
        "description": "Checks if the custom domain has correct DNS records configured.",
        "tags": ["Projects"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "DNS verification result" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Project not found" }
        }
      }
    },
...
  }
}